Clearer structure of DeckManager. Fixes small mistake in DeckManagerScreen & small refactoring.
