+++ src/screens/DeckManager/UpdateDeckScreen/styles.modules.tsx
@@ -1,6 +1,6 @@
 import {StyleSheet} from 'react-native';
 import { RFPercentage } from 'react-native-responsive-fontsize';
-import theme from './../../../assets/styles/theme.style';
+import theme from '../../../assets/styles/theme.style';
 
 const styles = StyleSheet.create({
   HeaderText: {
